Initial commit
This is some WIP for an implementation that is based on FlexBox
instantation.

It looks like the layout I've been trying to work on -- that of the
full, original MPC2000XL layout -- is more suitable for Grid rather
than FlexBox. So we'll commit what we have, because it's kinda cool,
and then we'll start working on an implementation based on Grid
instantiation.